made tests pass

+++ mockingbird/src/main/kotlin/com/leftindust/mediq/dao/DoctorDao.kt
@@ -7,5 +7,5 @@ import com.leftindust.mediq.extensions.CustomResult
 interface DoctorDao {
     suspend fun getByPatient(pid: Int, requester: MediqToken): CustomResult<List<Doctor>, OrmFailureReason>
     suspend fun getByVisit(vid: Long?, requester: MediqToken): CustomResult<Doctor, OrmFailureReason>
-    suspend fun getByDoctor(did: Int, requester: MediqToken): CustomResult<Doctor, OrmFailureReason>
+    suspend fun getByDoctor(did: Long, requester: MediqToken): CustomResult<Doctor, OrmFailureReason>
 }
